updated plugin elementor to 1.9.3

+++ wp-content/plugins/elementor/includes/template-library/sources/local.php
@@ -28,6 +28,8 @@ class Source_Local extends Source_Base {
 
 	private static $_template_types = [ 'page', 'section' ];
 
+	private $post_type_object;
+
 	/**
 	 * @static
 	 * @since 1.0.0
@@ -123,10 +125,18 @@ class Source_Local extends Source_Base {
 			'supports' => [ 'title', 'thumbnail', 'author', 'elementor' ],
 		];
 
-		register_post_type(
-			self::CPT,
-			apply_filters( 'elementor/template_library/sources/local/register_post_type_args', $args )
-		);
+		/**
+		 * Register template library post type args.
+		 *
+		 * Filters the post type arguments when registering elementor template library post type.
+		 *
+		 * @since 1.0.0
+		 *
+		 * @param array $args Arguments for registering a post type.
+		 */
+		$args = apply_filters( 'elementor/template_library/sources/local/register_post_type_args', $args );
+
+		$this->post_type_object = register_post_type( self::CPT, $args );
 
 		$args = [
 			'hierarchical' => false,
@@ -139,11 +149,18 @@ class Source_Local extends Source_Base {
 			'label' => _x( 'Type', 'Template Library', 'elementor' ),
 		];
 
-		register_taxonomy(
-			self::TAXONOMY_TYPE_SLUG,
-			self::CPT,
-			apply_filters( 'elementor/template_library/sources/local/register_taxonomy_args', $args )
-		);
+		/**
+		 * Register template library taxonomy args.
+		 *
+		 * Filters the taxonomy arguments when registering elementor template library taxonomy.
+		 *
+		 * @since 1.0.0
+		 *
+		 * @param array $args Arguments for registering a taxonomy.
+		 */
+		$args = apply_filters( 'elementor/template_library/sources/local/register_taxonomy_args', $args );
+
+		register_taxonomy( self::TAXONOMY_TYPE_SLUG, self::CPT, $args );
 	}
 
 	/**
@@ -154,8 +171,8 @@ class Source_Local extends Source_Base {
 		if ( current_user_can( 'manage_options' ) ) {
 			add_submenu_page(
 				Settings::PAGE_ID,
-				__( 'My Library', 'elementor' ),
-				__( 'My Library', 'elementor' ),
+				_x( 'My Library', 'Template Library', 'elementor' ),
+				_x( 'My Library', 'Template Library', 'elementor' ),
 				Editor::EDITING_CAPABILITY,
 				'edit.php?post_type=' . self::CPT
 			);
@@ -209,7 +226,7 @@ class Source_Local extends Source_Base {
 	}
 
 	/**
-	 * @since 1.0.1
+	 * @since 1.0.0
 	 * @access public
 	*/
 	public function save_item( $template_data ) {
@@ -217,9 +234,13 @@ class Source_Local extends Source_Base {
 			return new \WP_Error( 'save_error', 'Invalid template type `' . $template_data['type'] . '`' );
 		}
 
+		if ( ! current_user_can( $this->post_type_object->cap->edit_posts ) ) {
+			return new \WP_Error( 'save_error', __( 'Access Denied.', 'elementor' ) );
+		}
+
 		$template_id = wp_insert_post( [
 			'post_title' => ! empty( $template_data['title'] ) ? $template_data['title'] : __( '(no title)', 'elementor' ),
-			'post_status' => 'publish',
+			'post_status' => current_user_can( 'publish_posts' ) ? 'publish' : 'pending',
 			'post_type' => self::CPT,
 		] );
 
@@ -237,7 +258,28 @@ class Source_Local extends Source_Base {
 			SettingsManager::get_settings_managers( 'page' )->save_settings( $template_data['page_settings'], $template_id );
 		}
 
+		/**
+		 * After template library save.
+		 *
+		 * Fires after Elementor template library was saved.
+		 *
+		 * @since 1.0.1
+		 *
+		 * @param int   $template_id   The ID of the template.
+		 * @param array $template_data The template data.
+		 */
 		do_action( 'elementor/template-library/after_save_template', $template_id, $template_data );
+
+		/**
+		 * After template library update.
+		 *
+		 * Fires after Elementor template library was updated.
+		 *
+		 * @since 1.0.1
+		 *
+		 * @param int   $template_id   The ID of the template.
+		 * @param array $template_data The template data.
+		 */
 		do_action( 'elementor/template-library/after_update_template', $template_id, $template_data );
 
 		return $template_id;
@@ -248,8 +290,22 @@ class Source_Local extends Source_Base {
 	 * @access public
 	*/
 	public function update_item( $new_data ) {
+		if ( ! current_user_can( $this->post_type_object->cap->edit_post, $new_data['id'] ) ) {
+			return new \WP_Error( 'save_error', __( 'Access Denied.', 'elementor' ) );
+		}
+
 		Plugin::$instance->db->save_editor( $new_data['id'], $new_data['content'] );
 
+		/**
+		 * After template library update.
+		 *
+		 * Fires after Elementor template library was updated.
+		 *
+		 * @since 1.0.0
+		 *
+		 * @param int   $new_data_id The ID of the new template.
+		 * @param array $new_data    The new template data.
+		 */
 		do_action( 'elementor/template-library/after_update_template', $new_data['id'], $new_data );
 
 		return true;
@@ -269,22 +325,35 @@ class Source_Local extends Source_Base {
 
 		$page_settings = get_post_meta( $post->ID, PageSettingsManager::META_KEY, true );
 
+		$date = strtotime( $post->post_date );
+
 		$data = [
 			'template_id' => $post->ID,
 			'source' => $this->get_id(),
 			'type' => self::get_template_type( $post->ID ),
 			'title' => $post->post_title,
 			'thumbnail' => get_the_post_thumbnail_url( $post ),
-			'date' => mysql2date( get_option( 'date_format' ), $post->post_date ),
+			'date' => $date,
+			'human_date' => date_i18n( get_option( 'date_format' ), $date ),
 			'author' => $user->display_name,
 			'hasPageSettings' => ! empty( $page_settings ),
-			'categories' => [],
-			'keywords' => [],
+			'tags' => [],
 			'export_link' => $this->_get_export_link( $template_id ),
 			'url' => get_permalink( $post->ID ),
 		];
 
-		return apply_filters( 'elementor/template-library/get_template', $data );
+		/**
+		 * Get template library template.
+		 *
+		 * Filters the elementor template data when loading template library item.
+		 *
+		 * @since 1.0.0
+		 *
+		 * @param array $data Arguments for registering a taxonomy.
+		 */
+		$data = apply_filters( 'elementor/template-library/get_template', $data );
+
+		return $data;
 	}
 
 	/**
@@ -325,7 +394,11 @@ class Source_Local extends Source_Base {
 	 * @access public
 	*/
 	public function delete_template( $template_id ) {
-		wp_delete_post( $template_id, true );
+		if ( ! current_user_can( $this->post_type_object->cap->delete_post, $template_id ) ) {
+			return new \WP_Error( 'template_error', __( 'Access Denied.', 'elementor' ) );
+		}
+
+		return wp_delete_post( $template_id, true );
 	}
 
 	/**
@@ -437,15 +510,20 @@ class Source_Local extends Source_Base {
 
 		$items = [];
 
-		$zip = new \ZipArchive();
+		$file_extension = pathinfo( $_FILES['file']['name'], PATHINFO_EXTENSION );
 
-		/*
-		 * Check if file is a json or a .zip archive
-		 */
-		if ( true === $zip->open( $import_file ) ) {
+		if ( 'zip' === $file_extension ) {
+			if ( ! class_exists( '\ZipArchive' ) ) {
+				return new \WP_Error( 'zip_error', 'PHP Zip extension not loaded.' );
+			}
+
+			$zip = new \ZipArchive();
+
 			$wp_upload_dir = wp_upload_dir();
 
 			$temp_path = $wp_upload_dir['basedir'] . '/' . self::TEMP_FILES_DIR . '/' . uniqid();
+
+			$zip->open( $import_file );
 
 			$zip->extractTo( $temp_path );
 
@@ -495,15 +573,15 @@ class Source_Local extends Source_Base {
 		}
 		?>
 		<div id="elementor-hidden-area">
-			<a id="elementor-import-template-trigger" class="page-title-action"><?php _e( 'Import Templates', 'elementor' ); ?></a>
+			<a id="elementor-import-template-trigger" class="page-title-action"><?php esc_attr_e( 'Import Templates', 'elementor' ); ?></a>
 			<div id="elementor-import-template-area">
-				<div id="elementor-import-template-title"><?php _e( 'Choose an Elementor template JSON file or a .zip archive of Elementor templates, and add them to the list of templates available in your library.', 'elementor' ); ?></div>
+				<div id="elementor-import-template-title"><?php esc_attr_e( 'Choose an Elementor template JSON file or a .zip archive of Elementor templates, and add them to the list of templates available in your library.', 'elementor' ); ?></div>
 				<form id="elementor-import-template-form" method="post" action="<?php echo admin_url( 'admin-ajax.php' ); ?>" enctype="multipart/form-data">
 					<input type="hidden" name="action" value="elementor_import_template">
-					<input type="hidden" name="_nonce" value="<?php echo Plugin::$instance->editor->create_nonce(); ?>">
+					<input type="hidden" name="_nonce" value="<?php echo Plugin::$instance->editor->create_nonce( self::CPT ); ?>">
 					<fieldset id="elementor-import-template-form-inputs">
-						<input type="file" name="file" accept=".json,.zip,application/octet-stream,application/zip,application/x-zip,application/x-zip-compressed" required>
-						<input type="submit" class="button" value="<?php _e( 'Import Now', 'elementor' ); ?>">
+						<input type="file" name="file" accept=".json,application/json,.zip,application/octet-stream,application/zip,application/x-zip,application/x-zip-compressed" required>
+						<input type="submit" class="button" value="<?php esc_attr_e( 'Import Now', 'elementor' ); ?>">
 					</fieldset>
 				</form>
 			</div>
@@ -516,7 +594,7 @@ class Source_Local extends Source_Base {
 	 * @access public
 	*/
 	public function block_template_frontend() {
-		if ( is_singular( self::CPT ) && ! User::is_current_user_can_edit() ) {
+		if ( is_singular( self::CPT ) && ! current_user_can( 'edit_posts' ) ) {
 			wp_redirect( site_url(), 301 );
 			die;
 		}
@@ -527,10 +605,26 @@ class Source_Local extends Source_Base {
 	 * @access public
 	*/
 	public function is_template_supports_export( $template_id ) {
-		return apply_filters( 'elementor/template_library/is_template_supports_export', true, $template_id );
+		$export_support = true;
+
+		/**
+		 * Is template library supports export.
+		 *
+		 * Filters whether the template library supports export.
+		 *
+		 * @since 1.0.0
+		 *
+		 * @param bool $export_support Whether the template library supports export.
+		 *                             Default is true.
+		 * @param int  $template_id    Post ID.
+		 */
+		$export_support = apply_filters( 'elementor/template_library/is_template_supports_export', $export_support, $template_id );
+
+		return $export_support;
 	}
 
 	/**
+	 * @since 1.8.0
 	 * @access public
 	 */
 	public function remove_elementor_post_state_from_library( $post_states, $post ) {
@@ -549,7 +643,7 @@ class Source_Local extends Source_Base {
 			[
 				'action' => 'elementor_export_template',
 				'source' => $this->get_id(),
-				'_nonce' => Plugin::$instance->editor->create_nonce(),
+				'_nonce' => Plugin::$instance->editor->create_nonce( self::CPT ),
 				'template_id' => $template_id,
 			],
 			admin_url( 'admin-ajax.php' )
